+++ Anarchy/Message/Extensions.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace Discord
 {
@@ -54,7 +55,7 @@ namespace Discord
             var resp = client.HttpClient.Post($"/channels/{channelId}/typing");
 
             if (resp.Content.ReadAsStringAsync().Result.Contains("cooldown"))
-                throw new RateLimitException(client, resp.Deserialize<MessageRateLimit>().Cooldown);
+                throw new RateLimitException(client, resp.Deserialize<JObject>().GetValue("message_send_cooldown_ms").ToObject<uint>());
         }
     }
 }